requirements.txt
python-dotenv~=0.19.0
SQLAlchemy~=1.4.23
pydantic~=1.8.2
requests~=2.26.0
bs4~=0.0.1
beautifulsoup4~=4.10.0
telegram
psycopg2
